PLANET-4535: Split Timeline block scripts
