+++ assets/src/styles/blocks.scss
@@ -18,7 +18,6 @@
 @import "blocks/SplitTwoColumns";
 @import "blocks/Spreadsheet";
 @import "blocks/Submenu";
-@import "blocks/Timeline";
 @import "blocks/Cookies";
 
 // A/B test for Take Action Boxout
