+++ assets/src/styles/blocks/Timeline/TimelineStyle.scss
@@ -1,3 +1,7 @@
+@import "../../master-theme/assets/src/scss/base/variables";
+@import "../../master-theme/assets/src/scss/base/colors";
+@import "../../master-theme/assets/src/scss/base/mixins";
+
 .timeline-block {
   height: 600px;
